♻️ Refactor: useClickOutside hook 분리 #22
